+++ src/typings.ts
@@ -19,7 +19,7 @@ export type AlsoArray<T> = T | T[];
 /* <-- Compact Type Helpers --> */
 
 /** Generic for making any arbitrary function */
-export type AnyFunction<P extends any[] = any, R = any> = (...parameters: P) => R;
+export type AnyFunction<P extends any[] = any[], R = any> = (...parameters: P) => R;
 /** Allow either A or B but not both at the same time */
 export type Either<A, B> = Only<A, B> | Only<B, A>;
 export type Entries<T> = Array<{ [K in keyof T]: [keyof PickByValue<T, T[K]>, T[K]] }[keyof T]>;
@@ -40,6 +40,8 @@ export type Reverse<T extends any[]> = T extends [infer H, ...infer R] ? [...Rev
 /** Strict properties narrowing and remove Index Signatures */
 export type Strict<T> = { [P in keyof T as {} extends Record<P, any> ? never : P]: T[P] };
 export type Typify<T> = { [P in keyof T]: Typify<T[P]> };
+/** Any function that has exactly one parameter */
+export type UnaryFunction<P = any, R = any> = (parameter: P) => R;
 
 /* <-- Type Level Programming --> */
 
@@ -114,7 +116,7 @@ export type SingleProperty<T> = {
  * Specify tuple of `Size` with items of `T`
  */
 export type Tuple<
-	T,
 	Size extends number,
-	VirtualArray extends any[] = []
-> = VirtualArray['length'] extends Size ? VirtualArray : Tuple<T, Size, [T, ...VirtualArray]>;
+	T extends any[] = [],
+	Virtual extends any[] = []
+> = Virtual['length'] extends Size ? Virtual : Tuple<Size, T, [T, ...Virtual]>;